+++ plugins/permission/signers/index.ts
@@ -4,7 +4,8 @@ export {
 } from "./toECDSASigner.js"
 export {
     toWebAuthnSigner,
-    type WebAuthnModularSignerParams
+    type WebAuthnModularSignerParams,
+    WebAuthnSignerVersion
 } from "./toWebAuthnSigner.js"
 export { WebAuthnMode, toWebAuthnKey } from "@zerodev/webauthn-key"
 export { toSignerId } from "./utils/toSignerId.js"